ENHANCE: Remove legacy Config class and move hydra code into config.py
