+++ surround/surround/__init__.py
@@ -3,10 +3,9 @@ import pkg_resources
 from .run_modes import RunMode
 from .surround import Surround
 from .state import State
-from .config import Config, has_config
+from .config import SurroundConfig, BaseConfig, config, load_config
 from .stage import Stage, Estimator
 from .assembler import Assembler
 from .runners import Runner
-from .surround_config import SurroundConfig, BaseConfig, config, load_config
 
 __version__ = pkg_resources.get_distribution("surround").version
